tests/wpt/web-platform-tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.create2.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.imageData.create2.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.create2.basic</h1>
<p class="desc">createImageData(sw, sh) exists and returns something</p>


<script>
var t = async_test("createImageData(sw, sh) exists and returns something");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(async function() {

var canvas = new OffscreenCanvas(100, 50);
var ctx = canvas.getContext('2d');

await (function(canvas, ctx) {
_assertDifferent(ctx.createImageData(1, 1), null, "ctx.createImageData(1, 1)", "null");
})(canvas, ctx);
t.done();

});
</script>
